RDKOSS-159: Remove unrequired iptables-module packages
Reason for change: iptables has 100+ packages and among that RDK is using very few packages. With a basic experiment , it is identified that around 84 packages are not used and hence it is decoupled from RDEPENDS of iptables-modules to prevent installing it in rootfs.

RDKOSS-31: Remove unrequired util-linux packages

Reason for change: util-linux has 80+ packages and among that RDK is using very few packages. With a basic experiment , it is identified that around 54 packages are not used and hence it is decoupled from RRECOMMENDS of util-linux to prevent installing it in rootfs.

Signed-off-by: Arjun <arjun_daasuramdass@comcast.com>
(cherry picked from commit 4f0973987768a9f52895286401d7a4062cbe923a)
(cherry picked from commit 9c1af4370905ba5a329131421f2f6296b658f67f)