RDKOSS-295: Fix retriggering of "do_fetch" task for native modules
Reason for change: Previously, the "do_fetch" task for all native modules was retriggered for the image target, as opposed to other tasks that were sourced from the locked sstate cache. This issue arose due to the addition of the "do_populate_lic" task to all native recipes for the image target. This commit addresses the problem, ensuring consistent retrieval of tasks from the sstate cache.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>