RDKE-855: Include wayland-default-egl (#130)
* RDKE-855: Include wayland-default-egl

Reason for the change: The libwayland-egl.so has been removed from the Wayland module in OSS, with the assumption that the SoC vendor will provide an equivalent library as part of their platform-specific implementation.  the virtual/egl vendor module bundles libwayland-egl.so, effectively resolving dependencies for other layer modules that rely on libwayland-egl.so.

However, it has been found that some SoC platforms still rely on the libwayland-egl.so library being provided by the Wayland module. Including this library back into the Wayland module would break the build for other SoC platforms that have already providing this lib.

To address this issue, The wayland-default-egl recipe added to the OSS layer, will now provide the default libwayland-egl.so library which can be used on demand, until all platforms have transitioned to the new OSS consumption method, which supports platform specific OSS custamization.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>

* Update wayland-default-egl_1.20.0.bb

---------

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>
Co-authored-by: mselva006c <mani_selvaraj@comcast.com>