RDKOSS-490: Define OSS_LAYER_VERSION
Reason for the change - The meta-oss-reference-release repository has been excluded from the manifest to align with the new OSS consumption model. As a result, the OSS_LAYER_VERSION parameter defined within that repository is no longer available to the build. Move the OSS_LAYER_VERSION parameter to the reference layer