Feature/oss rel 4.7.5.2 (#224)
* RDKOSS-490: Define OSS_LAYER_VERSION (#222)

Reason for the change - The meta-oss-reference-release repository has been excluded from the manifest to align with the new OSS consumption model. As a result, the OSS_LAYER_VERSION parameter defined within that repository is no longer available to the build. Move the OSS_LAYER_VERSION parameter to the reference layer

* RDKE-942: Update OSS version to 4.7.5.2

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>

* Update package_revisions_oss.inc

---------

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>
Co-authored-by: mselva006c <mani_selvaraj@comcast.com>